+++ system/src/Grav/Common/Flex/Pages/PageObject.php
@@ -19,6 +19,7 @@ use Grav\Common\Flex\Pages\Traits\PageRoutableTrait;
 use Grav\Common\Flex\Pages\Traits\PageTranslateTrait;
 use Grav\Common\Page\Pages;
 use Grav\Common\Utils;
+use Grav\Framework\Filesystem\Filesystem;
 use Grav\Framework\Flex\FlexObject;
 use Grav\Framework\Flex\Pages\FlexPageObject;
 use Grav\Framework\Route\Route;
@@ -108,7 +109,8 @@ class PageObject extends FlexPageObject
                 // TODO: this should not be template!
                 return $this->getProperty('template');
             case 'route':
-                $key = dirname($this->hasKey() ? '/' . $this->getKey() : '/');
+                $filesystem = Filesystem::getInstance(false);
+                $key = $filesystem->dirname($this->hasKey() ? '/' . $this->getKey() : '/');
                 return $key !== '/' ? $key : null;
             case 'full_route':
                 return $this->hasKey() ? '/' . $this->getKey() : '';
@@ -151,8 +153,10 @@ class PageObject extends FlexPageObject
      */
     protected function reorderSiblings(array $ordering)
     {
+        $filesystem = Filesystem::getInstance(false);
+
         $storageKey = $this->getStorageKey();
-        $oldParentKey = ltrim(dirname("/$storageKey"), '/');
+        $oldParentKey = ltrim($filesystem->dirname("/$storageKey"), '/');
         $newParentKey = $this->getProperty('parent_key');
 
         $slug = basename($this->getKey());